added the ability to have columns and rows within lines, and upadted all code to match the behavior that existed before they were added. we still need to work out the css and try adding columns and rows in the data for the program to render from

+++ build.ps1
@@ -38,7 +38,9 @@ else {
     else {
         if ($release.IsPresent -or $minify.IsPresent) {
             Write-Host "    Removing old JavaScript..."
-            Remove-Item -Recurse -Force ./js
+            if (Test-Path ./js -PathType Container) {
+                Remove-Item -Recurse -Force ./js
+            }
         }
         
         Write-Host "    Compiling TypeScript into JavaScript..."
